requirements.txt
aiohttp
asyncio
python-dotenv
python-telegram-bot
